+++ BALSAMIC/utils/utils.py
@@ -1,0 +1,6 @@
+"""Helper functions."""
+
+
+def remove_unnecessary_spaces(string: str) -> str:
+    """Return a string removing unnecessary empty spaces."""
+    return " ".join(string.split())